Problem: Partial<Weather> is a hassle.
It's going to be a problem for blending two weather tiles together.

Solution: Restrict it to the Picker.
Everywhere else is a legal Weather or null.

This attempt uses a local Partial<Weather> state that is kept in sync
with the global weather (via weather prop).

The use of useEffect() is not ideal, as it causes a double render.

````
  useEffect(() => {
    _setLocalWeather(weather || {});
  }, [weather]);
```